feat: Extract the useFetchModelId to logicHooks.ts
